Relocate all Bridgetown Roda logic to new `bridgetown_stack` plugin
- Also catch syntax errors in dev watchers
