+++ bridgetown-routes/test/helper.rb
@@ -3,5 +3,6 @@
 require_relative "../../bridgetown-core/test/helper"
 require "bridgetown-builder"
 require "rack"
+require "rack/test"
 
 Bridgetown.begin!
